requirements.txt
astroid==3.3.8
autopep8==2.3.2
blinker==1.9.0
click==8.1.8
dill==0.3.9
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
isort==6.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
loguru==0.7.3
MarkupSafe==3.0.2
mccabe==0.7.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pylint==3.3.4
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.8.2
python-decouple==3.8
SQLAlchemy==2.0.37
tomlkit==0.13.2
typing_extensions==4.12.2
Werkzeug==3.1.3